# user-service/openapi/user-service.openapi.yaml
openapi: 3.1.0
info:
  title: User Service API
  version: 0.1.0
  description: Auth0-protected identity, self-service, and user administration API.
servers:
  - url: http://localhost:8080
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /admin/users:
    get:
      tags: [user-admin]
      summary: List users
      operationId: listUsers
      x-required-permission: read:users
      security:
        - bearerAuth: ["read:users"]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/StartQuery"
        - $ref: "#/components/parameters/EndQuery"
      responses:
        "200":
          description: Auth0 user search result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0ListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}:
    get:
      tags: [user-admin]
      summary: Read user
      operationId: readUser
      x-required-permission: read:users
      security:
        - bearerAuth: ["read:users"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: Auth0 user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/attributes:
    patch:
      tags: [user-admin]
      summary: Update user attributes
      operationId: updateUserAttributes
      x-required-permission: update:users
      security:
        - bearerAuth: ["update:users"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserAttributeUpdateRequest"
      responses:
        "200":
          description: Updated Auth0 user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/disable:
    post:
      tags: [user-admin]
      summary: Disable user
      operationId: disableUser
      x-required-permission: disable:users
      security:
        - bearerAuth: ["disable:users"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDisableRequest"
      responses:
        "200":
          description: Auth0 user blocked with user-service workflow metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/delete-workflow:
    post:
      tags: [user-admin]
      summary: Start user delete workflow
      operationId: startUserDeleteWorkflow
      x-required-permission: delete:users
      security:
        - bearerAuth: ["delete:users"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserDeleteWorkflowRequest"
      responses:
        "200":
          description: Auth0 user blocked and marked for delayed delete workflow.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0User"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/password-reset:
    post:
      tags: [user-admin]
      summary: Create password reset ticket
      operationId: resetUserPassword
      x-required-permission: reset:passwords
      security:
        - bearerAuth: ["reset:passwords"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "200":
          description: Auth0 password change ticket.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/mfa/reset:
    post:
      tags: [user-admin]
      summary: Reset user MFA
      operationId: resetUserMfa
      x-required-permission: reset:mfa
      security:
        - bearerAuth: ["reset:mfa"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminMfaResetRequest"
      responses:
        "200":
          description: Auth0 MFA reset result.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/groups:
    get:
      tags: [user-admin]
      summary: List groups
      operationId: listGroups
      x-required-permission: read:groups
      security:
        - bearerAuth: ["read:groups"]
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/Query"
        - $ref: "#/components/parameters/StartQuery"
        - $ref: "#/components/parameters/EndQuery"
      responses:
        "200":
          description: Auth0 role search result. Groups are represented as Auth0 Roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0ListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    post:
      tags: [user-admin]
      summary: Create group
      operationId: createGroup
      x-required-permission: create:groups
      security:
        - bearerAuth: ["create:groups"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupCreateRequest"
      responses:
        "201":
          description: Created Auth0 Role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/groups/{group_id}:
    get:
      tags: [user-admin]
      summary: Read group
      operationId: readGroup
      x-required-permission: read:groups
      security:
        - bearerAuth: ["read:groups"]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "200":
          description: Auth0 Role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    patch:
      tags: [user-admin]
      summary: Update group
      operationId: updateGroup
      x-required-permission: update:groups
      security:
        - bearerAuth: ["update:groups"]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdateRequest"
      responses:
        "200":
          description: Updated Auth0 Role.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    delete:
      tags: [user-admin]
      summary: Delete group
      operationId: deleteGroup
      x-required-permission: delete:groups
      security:
        - bearerAuth: ["delete:groups"]
      parameters:
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: Group deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/groups/{group_id}/users:
    get:
      tags: [user-admin]
      summary: List group users
      operationId: listGroupUsers
      x-required-permission: read:groups
      security:
        - bearerAuth: ["read:groups"]
      parameters:
        - $ref: "#/components/parameters/GroupId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: Auth0 Role users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0ListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/permissions:
    post:
      tags: [user-admin]
      summary: Create API permission
      operationId: createPermission
      x-required-permission: create:permissions
      security:
        - bearerAuth: ["create:permissions"]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionCreateRequest"
      responses:
        "201":
          description: Updated Auth0 resource server with API scopes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0ResourceServer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/permissions/{permission}:
    patch:
      tags: [user-admin]
      summary: Update API permission
      operationId: updatePermission
      x-required-permission: update:permissions
      security:
        - bearerAuth: ["update:permissions"]
      parameters:
        - name: permission
          in: path
          required: true
          schema:
            type: string
          example: read:orders
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PermissionUpdateRequest"
      responses:
        "200":
          description: Updated Auth0 resource server with API scopes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0ResourceServer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    delete:
      tags: [user-admin]
      summary: Delete API permission
      operationId: deletePermission
      x-required-permission: delete:permissions
      security:
        - bearerAuth: ["delete:permissions"]
      parameters:
        - name: permission
          in: path
          required: true
          schema:
            type: string
          example: read:orders
      responses:
        "204":
          description: Permission removed from Auth0 resource server.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/permissions:
    get:
      tags: [user-admin]
      summary: List user permissions
      operationId: listUserPermissions
      x-required-permission: read:permissions
      security:
        - bearerAuth: ["read:permissions"]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - $ref: "#/components/parameters/UseCache"
      responses:
        "200":
          description: Auth0 user permissions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0UserPermissionsResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    post:
      tags: [user-admin]
      summary: Assign permissions to user
      operationId: assignPermissionsToUser
      x-required-permission: assign:permissions
      security:
        - bearerAuth: ["assign:permissions"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPermissionAssignmentRequest"
      responses:
        "204":
          description: Permissions assigned to user.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    delete:
      tags: [user-admin]
      summary: Remove permissions from user
      operationId: removePermissionsFromUser
      x-required-permission: unassign:permissions
      security:
        - bearerAuth: ["unassign:permissions"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPermissionAssignmentRequest"
      responses:
        "204":
          description: Permissions removed from user.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/groups:
    get:
      tags: [user-admin]
      summary: List user groups
      operationId: listUserGroups
      x-required-permission: read:groups
      security:
        - bearerAuth: ["read:groups"]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
      responses:
        "200":
          description: User Auth0 Roles.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auth0ListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
    post:
      tags: [user-admin]
      summary: Add user to group
      operationId: addUserToGroup
      x-required-permission: update:groups
      security:
        - bearerAuth: ["update:groups"]
      parameters:
        - $ref: "#/components/parameters/UserId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupAssignmentRequest"
      responses:
        "204":
          description: User assigned to group.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
  /admin/users/{user_id}/groups/{group_id}:
    delete:
      tags: [user-admin]
      summary: Remove user from group
      operationId: removeUserFromGroup
      x-required-permission: delete:groups
      security:
        - bearerAuth: ["delete:groups"]
      parameters:
        - $ref: "#/components/parameters/UserId"
        - $ref: "#/components/parameters/GroupId"
      responses:
        "204":
          description: User removed from group.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "502":
          $ref: "#/components/responses/Auth0Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 access token with route-specific permissions in the permissions claim.
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
    Query:
      name: query
      in: query
      schema:
        type: string
      description: Primary search fragment.
    StartQuery:
      name: start_query
      in: query
      schema:
        type: string
      description: Optional search fragment used as the start bound.
    EndQuery:
      name: end_query
      in: query
      schema:
        type: string
      description: Optional search fragment used as the end bound.
...
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
      example: auth0|USER_ID
    GroupId:
      name: group_id
      in: path
      required: true
      schema:
        type: string
      example: rol_GROUP_ID
...
  responses:
    Unauthorized:
      description: Missing, malformed, or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/UnauthorizedError"
    Forbidden:
      description: Bearer token is valid but missing one or more required permissions.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ForbiddenError"
    Auth0Error:
      description: Auth0 upstream error translated by the service.
      content:
        application/json:
          schema:
            type: object
            additionalProperties: true
  schemas:
...
    UserAttributeUpdateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        blocked:
          type: boolean
        email_verified:
          type: boolean
        user_metadata:
          type: object
          additionalProperties: true
        app_metadata:
          type: object
          additionalProperties: true
    UserDisableRequest:
      type: object
      properties:
        reason:
          type: string
          description: Optional admin reason stored in Auth0 app_metadata.user_service_workflow.
          example: Suspicious activity review
    UserDeleteWorkflowRequest:
      type: object
      properties:
        reason:
          type: string
          description: Optional admin reason stored in Auth0 app_metadata.user_service_workflow.
          example: Customer requested account closure
        retention_days:
          type: integer
          minimum: 1
          description: Optional retention period before external deletion processing may run.
          example: 30
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
        connection_id:
          type: string
        result_url:
          type: string
    AdminMfaResetRequest:
      type: object
      properties:
        authentication_method_id:
          type: string
          description: Optional Auth0 authentication method id. Omit to reset all methods.
...
    GroupCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
    GroupUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
...
    Auth0User:
      type: object
      additionalProperties: true
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
    Auth0Role:
      type: object
      additionalProperties: true
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
...
    Auth0ListResponse:
      type: object
      additionalProperties: true
      properties:
        start:
          type: integer
        limit:
          type: integer
        length:
          type: integer
        total:
          type: integer
        users:
          type: array
          items:
            $ref: "#/components/schemas/Auth0User"
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Auth0Role"
...